тесты все + gcov

+++ src/s21_from_float_to_decimal.c
@@ -8,7 +8,7 @@ int s21_from_float_to_decimal(float src, s21_decimal *dst){
     dst -> bits[0] = dst -> bits[1] = dst -> bits[2] = dst -> bits[3] = 0;
 
     int sign = (src < 0);
-    float src_abs = fabsf(src); // для большей точности
+    double src_abs = fabs((double)src); // для большей точности
 
     if (src_abs > 0 && src_abs < 1e-28){
         flag = 1;
